+++ cmd/hidden_lake/service/internal/config/editor_test.go
@@ -44,6 +44,7 @@ var (
 
 func (p *tsConfig) GetSettings() IConfigSettings              { return nil }
 func (p *tsConfig) GetLogging() logger.ILogging               { return nil }
+func (p *tsConfig) GetShare() bool                            { return false }
 func (p *tsConfig) GetAddress() IAddress                      { return nil }
 func (p *tsConfig) GetNetworkKey() string                     { return "" }
 func (p *tsConfig) GetConnections() []string                  { return nil }